foundry_evm::executors
Module ITest

Module containing a contract’s types and functions.

interface ITest {
    function setUp() external;
    function failed() external view returns (bool failed);
    function beforeTestSetup(bytes4 testSelector) public view returns (bytes[] memory beforeTestCalldata);
}

Structs

beforeTestSetupCall

Function with signature beforeTestSetup(bytes4) and selector 0x49b07678.

beforeTestSetupReturn

Container type for the return parameters of the beforeTestSetup(bytes4) function.

failedCall

Function with signature failed() and selector 0xba414fa6.

failedReturn

Container type for the return parameters of the failed() function.

setUpCall

Function with signature setUp() and selector 0x0a9254e4.

setUpReturn

Container type for the return parameters of the setUp() function.

Enums

ITestCalls

Container for all the ITest function calls.