foundry_evm::executors
Module fuzz

Modules

types 🔒

Structs

CaseOutcome

Returned by a single fuzz in the case of a successful run

CounterExampleOutcome

Returned by a single fuzz when a counterexample has been discovered

FuzzTestData

Contains data collected during fuzz test runs.

FuzzedExecutor

Wrapper around an Executor which provides fuzzing support using [proptest].

Enums

FuzzOutcome

Outcome of a single fuzz