Module foundry_evm::executors::fuzz::types

Structs

CaseOutcome

Returned by a single fuzz in the case of a successful run

CounterExampleOutcome

Returned by a single fuzz when a counterexample has been discovered

Enums

FuzzOutcome

Outcome of a single fuzz