foundry_evm::executors::invariant

Module IInvariantTest

Module containing a contract’s types and functions.

interface IInvariantTest {
    struct FuzzSelector { address addr; bytes4[] selectors; }
    struct FuzzArtifactSelector { string artifact; bytes4[] selectors; }
    struct FuzzInterface { address addr; string[] artifacts; }
    function afterInvariant() external;
    function excludeArtifacts() public view returns (string[] memory excludedArtifacts);
    function excludeContracts() public view returns (address[] memory excludedContracts);
    function excludeSelectors() public view returns (FuzzSelector[] memory excludedSelectors);
    function excludeSenders() public view returns (address[] memory excludedSenders);
    function targetArtifacts() public view returns (string[] memory targetedArtifacts);
    function targetArtifactSelectors() public view returns (FuzzArtifactSelector[] memory targetedArtifactSelectors);
    function targetContracts() public view returns (address[] memory targetedContracts);
    function targetSelectors() public view returns (FuzzSelector[] memory targetedSelectors);
    function targetSenders() public view returns (address[] memory targetedSenders);
    function targetInterfaces() public view returns (FuzzInterface[] memory targetedInterfaces);
}

Structs

FuzzArtifactSelector

FuzzInterface

FuzzSelector

afterInvariantCall

Function with signature afterInvariant() and selector 0x93969ddf.

afterInvariantReturn

Container type for the return parameters of the afterInvariant() function.

excludeArtifactsCall

Function with signature excludeArtifacts() and selector 0xb5508aa9.

excludeArtifactsReturn

Container type for the return parameters of the excludeArtifacts() function.

excludeContractsCall

Function with signature excludeContracts() and selector 0xe20c9f71.

excludeContractsReturn

Container type for the return parameters of the excludeContracts() function.

excludeSelectorsCall

Function with signature excludeSelectors() and selector 0xb0464fdc.

excludeSelectorsReturn

Container type for the return parameters of the excludeSelectors() function.

excludeSendersCall

Function with signature excludeSenders() and selector 0x1ed7831c.

excludeSendersReturn

Container type for the return parameters of the excludeSenders() function.

targetArtifactSelectorsCall

Function with signature targetArtifactSelectors() and selector 0x66d9a9a0.

targetArtifactSelectorsReturn

Container type for the return parameters of the targetArtifactSelectors() function.

targetArtifactsCall

Function with signature targetArtifacts() and selector 0x85226c81.

targetArtifactsReturn

Container type for the return parameters of the targetArtifacts() function.

targetContractsCall

Function with signature targetContracts() and selector 0x3f7286f4.

targetContractsReturn

Container type for the return parameters of the targetContracts() function.

targetInterfacesCall

Function with signature targetInterfaces() and selector 0x2ade3880.

targetInterfacesReturn

Container type for the return parameters of the targetInterfaces() function.

targetSelectorsCall

Function with signature targetSelectors() and selector 0x916a17c6.

targetSelectorsReturn

Container type for the return parameters of the targetSelectors() function.

targetSendersCall

Function with signature targetSenders() and selector 0x3e5e3c23.

targetSendersReturn

Container type for the return parameters of the targetSenders() function.

Enums

IInvariantTestCalls

Container for all the IInvariantTest function calls.