foundry_evm::executors::invariant
Module result

Structs

InvariantFuzzTestResult

The outcome of an invariant fuzz test

RichInvariantResults 🔒

Enriched results of an invariant run check.

Functions

assert_after_invariant 🔒

Given the executor state, asserts conditions within afterInvariant function. If call fails then the invariant test is considered failed.

assert_invariants 🔒

Given the executor state, asserts that no invariant has been broken. Otherwise, it fills the external invariant_failures.failed_invariant map and returns a generic error. Either returns the call result if successful, or nothing if there was an error.

can_continue 🔒

Returns if invariant test can continue and last successful call result of the invariant test function (if it can continue).