foundry_evm::executors::invariant::result
Function can_continue

pub(crate) fn can_continue(
    invariant_contract: &InvariantContract<'_>,
    invariant_test: &InvariantTest,
    invariant_run: &mut InvariantTestRun,
    invariant_config: &InvariantConfig,
    call_result: RawCallResult,
    state_changeset: &StateChangeset,
) -> Result<RichInvariantResults>

Returns if invariant test can continue and last successful call result of the invariant test function (if it can continue).