foundry_evm::executors::invariant

Module shrink

Structs

CallSequenceShrinker 🔒

Shrinker for a call sequence failure. Iterates sequence call sequence top down and removes calls one by one. If the failure is still reproducible with removed call then moves to the next one. If the failure is not reproducible then restore removed call and moves to next one.

Shrink 🔒

Functions

check_sequence

Checks if the given call sequence breaks the invariant.

shrink_sequence 🔒

Shrinks the failure case to its smallest sequence of calls.