Module foundry_evm::inspectors

EVM inspectors.

Re-exports

pub use foundry_cheatcodes as cheatcodes;

Modules

chisel_state 🔒

debugger 🔒

logs 🔒

stack 🔒

Structs

AccessListInspector

An [Inspector] that collects touched accounts and storage slots.

Cheatcodes

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

CheatsConfig

Additional, configurable context the Cheatcodes inspector has access to

ChiselState

An inspector for Chisel

CoverageCollector

Debugger

An inspector that collects debug nodes on every step of the interpreter.

Fuzzer

An inspector that can fuzz and collect data for that effect.

InspectorData

The collected results of InspectorStack.

InspectorStack

An inspector that calls multiple inspectors in sequence.

InspectorStackBuilder

LogCollector

An inspector that collects logs during execution.

TracingInspector

An inspector that collects call traces.

TracingInspectorConfig

Gives guidance to the TracingInspector.

Enums

StackSnapshotType

How much of the stack to record. Nothing, just the items pushed, or the full stack