foundry_evm

Module utils

Structs

IcPcMap

Map from instruction counter to program counter.

Instruction

Represents a single instruction consisting of the opcode and its immediate data.

PcIcMap

Maps from program counter to instruction counter.

Functions

apply_chain_and_block_specific_env_changes

Depending on the configured chain id and block number this should apply any specific changes

configure_tx_env

Configures the env for the given RPC transaction. Accounts for an impersonated transaction by resetting the env.tx.caller field to tx.from.

configure_tx_req_env

Configures the env for the given RPC transaction request. impersonated_from is the address of the impersonated account. This helps account for an impersonated transaction by resetting the env.tx.caller field to impersonated_from.

create2_handler_register

Used for routing certain CREATE2 invocations through CREATE2_DEPLOYER.

decode_instructions

Decodes raw opcode bytes into Instructions.

gas_used

Get the gas used, accounting for refunds

get_function

Given an ABI and selector, it tries to find the respective function.

new_evm_with_existing_context

new_evm_with_inspector

Creates a new EVM with the given inspector.

odyssey_handler_register

Adds Odyssey P256 precompile to the list of loaded precompiles.

Type Aliases

StateChangeset

EVM State is a mapping from addresses to accounts.