Crate foundry_evm

foundry-evm

Main Foundry EVM backend abstractions.

Re-exports

pub use foundry_evm_coverage as coverage;

pub use foundry_evm_fuzz as fuzz;

pub use foundry_evm_traces as traces;

Modules

backend

Foundry’s main executor backend abstraction and implementation.

constants

decode

Various utilities to decode test results.

executors

EVM executor abstractions, which can execute calls.

fork

inspectors

EVM inspectors.

opts

utils

Macros

call_inspectors

Helper macro to call the same method on multiple inspectors without resorting to dynamic dispatch.

Structs

Env

Helper container type for EvmEnv and [TxEnv].

EnvMut

Helper struct with mutable references to the block and cfg environments.

EvmEnv

Container type that holds both the configuration and block environment for EVM execution.

Traits

InspectorExt

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.