foundry_evm

Module executors

EVM executor abstractions, which can execute calls.

Used for running tests, scripts, and interacting with the inner backend which holds the state.

Re-exports

pub use fuzz::FuzzedExecutor;

pub use invariant::InvariantExecutor;

Modules

ITest

Module containing a contract’s types and functions.

builder 🔒

corpus 🔒

fuzz

invariant

trace 🔒

Structs

CallResult

The result of a call.

DeployResult

The result of a deployment.

ExecutionErr

Represents the context after an execution error occurred.

Executor

EVM executor.

ExecutorBuilder

The builder that allows to configure an evm Executor which a stack of optional [revm::Inspector]s, such as Cheatcodes.

FailFast

Helper struct to enable fail fast behavior: when one test fails, all other tests stop early.

FuzzTestTimer

Timer for a fuzz test.

RawCallResult

The result of a raw call.

TracingExecutor

A default executor with tracing enabled

Enums

EvmError

Constants

DURATION_BETWEEN_METRICS_REPORT 🔒

Functions

convert_executed_result 🔒

Converts the data aggregated in the inspector and call to a RawCallResult