Module foundry_evm::executors

EVM executor abstractions, which can execute calls.

Used for running tests, scripts, and interacting with the inner backend which holds the state.

Re-exports

pub use fuzz::FuzzedExecutor;

pub use invariant::InvariantExecutor;

Modules

ITest

Module containing a contract’s types and functions.

builder 🔒

fuzz

invariant

trace 🔒

Structs

CallResult

The result of a call.

DeployResult

The result of a deployment.

ExecutionErr

Represents the context after an execution error occurred.

Executor

EVM executor.

ExecutorBuilder

The builder that allows to configure an evm Executor which a stack of optional [revm::Inspector]s, such as Cheatcodes.

RawCallResult

The result of a raw call.

TracingExecutor

A default executor with tracing enabled

Enums

EvmError

Functions

convert_executed_result 🔒

Converts the data aggregated in the inspector and call to a RawCallResult