foundry_evm::executors::invariant

Module replay

Functions

replay_error

Replays the error case, shrinks the failing sequence and collects all necessary traces.

replay_run

Replays a call sequence for collecting logs and traces. Returns counterexample to be used when the call sequence is a failed scenario.

set_up_inner_replay 🔒

Sets up the calls generated by the internal fuzzer, if they exist.