Module foundry_evm::executors::invariant::shrink

Structs

CallSequenceShrinker 🔒

Shrinker for a call sequence failure. Iterates sequence call sequence top down and removes calls one by one. If the failure is still reproducible with removed call then moves to the next one. If the failure is not reproducible then restore removed call and moves to next one.

Shrink 🔒

Functions

check_sequence

Checks if the given call sequence breaks the invariant. Used in shrinking phase for checking candidate sequences and in replay failures phase to test persisted failures. Returns the result of invariant check (and afterInvariant call if needed) and if sequence was entirely applied.

shrink_sequence 🔒

Shrinks the failure case to its smallest sequence of calls.