foundry_evm

Module inspectors

EVM inspectors.

Re-exports

pub use foundry_cheatcodes as cheatcodes;

Modules

chisel_state 🔒

custom_printer 🔒

Custom print inspector, it has step level information of execution. It is a great tool if some debugging is needed.

logs 🔒

revert_diagnostic 🔒

script 🔒

stack 🔒

Structs

AccessListInspector

An [Inspector] that collects touched accounts and storage slots.

Cheatcodes

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

CheatsConfig

Additional, configurable context the Cheatcodes inspector has access to

ChiselState

An inspector for Chisel

CustomPrintTracer

Custom print [Inspector], it has step level information of execution.

Fuzzer

An inspector that can fuzz and collect data for that effect.

InspectorData

The collected results of InspectorStack.

InspectorStack

An inspector that calls multiple inspectors in sequence.

InspectorStackBuilder

LineCoverageCollector

Inspector implementation for collecting coverage information.

LogCollector

An inspector that collects logs during execution.

RevertDiagnostic

An inspector that tracks call context to enhances revert diagnostics. Useful for understanding reverts that are not linked to custom errors or revert strings.

ScriptExecutionInspector

An inspector that enforces certain rules during script execution.

TracingInspector

An inspector that collects call traces.

TracingInspectorConfig

Gives guidance to the TracingInspector.

Enums

StackSnapshotType

How much of the stack to record. Nothing, just the items pushed, or the full stack