Module foundry_evm::utils

Structs

IcPcMap

Map from instruction counter to program counter.

PcIcMap

Maps from program counter to instruction counter.

Functions

apply_chain_and_block_specific_env_changes

Depending on the configured chain id and block number this should apply any specific changes

configure_tx_env

Configures the env for the transaction

create2_handler_register

Used for routing certain CREATE2 invocations through DEFAULT_CREATE2_DEPLOYER.

gas_used

Get the gas used, accounting for refunds

get_function

Given an ABI and selector, it tries to find the respective function.

new_evm_with_existing_context

new_evm_with_inspector

Creates a new EVM with the given inspector.

new_evm_with_inspector_ref

Creates a new EVM with the given inspector and wraps the database in a WrapDatabaseRef.

Type Aliases

StateChangeset

EVM State is a mapping from addresses to accounts.